+++ app/Models/User.php
@@ -217,7 +217,7 @@ class User extends Authenticatable implements HasLocalePreference
     }
 
     /**
-     * Scope a query to only get users that have a name like the passed one.
+     * Scope a query to only get users that have a name/email like the passed one.
      *
      * The name gets split up by the whitespaces and each part will be searched
      * in the corresponding name fields.