+++ CHANGELOG.md
@@ -13,15 +13,20 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Permission restrictions to prevent non-superusers from assigning the superuser role ([#1651])
 - Environment variable for configuring restricted permissions that cannot be assigned to non-superuser roles ([#1651])
 - Display raw permission names in the admin interface ([#1651])
+- Visual tests with [Happo.io](https://happo.io) ([#1600])
+- Artisan command for provisioning via JSON file ([#1636], [#1678])
 
 ### Changed
 
 - Real-time input validation on create superuser command ([#1651])
+- Error handling in room statistics ([#1535], [#1600])
+- Error handling in room attendance ([#1535], [#1600])
 
 ### Fixed
 
 - Wrong error message shown for 422 errors when verifying email ([#1744], [#1758])
 - Broken banner link style 'warning' ([#1759], [#1760])
+- Inconsistent select/multiselect loading states ([#1772])
 
 ## [v4.2.0] - 2025-01-06
 
@@ -310,6 +315,7 @@ You can find the changelog for older versions there [here](https://github.com/TH
 [#1493]: https://github.com/THM-Health/PILOS/issues/1493
 [#1495]: https://github.com/THM-Health/PILOS/pull/1495
 [#1523]: https://github.com/THM-Health/PILOS/issues/1523
+[#1535]: https://github.com/THM-Health/PILOS/issues/1535
 [#1536]: https://github.com/THM-Health/PILOS/issues/1536
 [#1537]: https://github.com/THM-Health/PILOS/pull/1537
 [#1548]: https://github.com/THM-Health/PILOS/pull/1548
@@ -320,6 +326,7 @@ You can find the changelog for older versions there [here](https://github.com/TH
 [#1565]: https://github.com/THM-Health/PILOS/pull/1565
 [#1569]: https://github.com/THM-Health/PILOS/pull/1569
 [#1595]: https://github.com/THM-Health/PILOS/pull/1595
+[#1600]: https://github.com/THM-Health/PILOS/pull/1600
 [#1604]: https://github.com/THM-Health/PILOS/pull/1604
 [#1607]: https://github.com/THM-Health/PILOS/issues/1607
 [#1608]: https://github.com/THM-Health/PILOS/pull/1608
@@ -336,6 +343,7 @@ You can find the changelog for older versions there [here](https://github.com/TH
 [#1758]: https://github.com/THM-Health/PILOS/pull/1758
 [#1759]: https://github.com/THM-Health/PILOS/issues/1759
 [#1760]: https://github.com/THM-Health/PILOS/pull/1760
+[#1772]: https://github.com/THM-Health/PILOS/pull/1772
 [unreleased]: https://github.com/THM-Health/PILOS/compare/v4.2.0...develop
 [v3.0.0]: https://github.com/THM-Health/PILOS/releases/tag/v3.0.0
 [v3.0.1]: https://github.com/THM-Health/PILOS/releases/tag/v3.0.1